Add to Package aka Shopping Cart (#2955)

+++ catalog/app/components/FileEditor/FileEditor.tsx
@@ -3,6 +3,8 @@ import * as RRDom from 'react-router-dom'
 
 import * as PreviewUtils from 'components/Preview/loaders/utils'
 import PreviewDisplay from 'components/Preview/Display'
+import type * as Model from 'model'
+import * as AddToPackage from 'containers/AddToPackage'
 import AsyncResult from 'utils/AsyncResult'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import parseSearch from 'utils/parseSearch'
@@ -14,19 +16,25 @@ import { detect, loadMode, useWriteData } from './loader'
 import { EditorInputType } from './types'
 
 function useRedirect() {
+  const addToPackage = AddToPackage.use()
   const history = RRDom.useHistory()
   const { urls } = NamedRoutes.use()
   const location = RRDom.useLocation()
-  const { next } = parseSearch(location.search, true)
+  const { add, next } = parseSearch(location.search, true)
   return React.useCallback(
-    ({ bucket, key, version }) =>
-      history.push(next || urls.bucketFile(bucket, key, { version })),
-    [history, next, urls],
+    ({ bucket, key, size, version }: Model.S3File) => {
+      if (add) {
+        addToPackage.append({ bucket, key, size, version })
+      }
+      history.push(next || urls.bucketFile(bucket, key, { version }))
+    },
+    [history, next, addToPackage, add, urls],
   )
 }
 
 interface EditorState {
   editing: boolean
+  error: Error | null
   onCancel: () => void
   onChange: (value: string) => void
   onEdit: () => void
@@ -40,6 +48,7 @@ export function useState(handle: S3HandleBase): EditorState {
   const type = React.useMemo(() => detect(handle.key), [handle.key])
   const location = RRDom.useLocation()
   const { edit } = parseSearch(location.search, true)
+  const [error, setError] = React.useState<Error | null>(null)
   const [value, setValue] = React.useState<string | undefined>()
   const [editing, setEditing] = React.useState<boolean>(!!edit)
   const [saving, setSaving] = React.useState<boolean>(false)
@@ -47,16 +56,25 @@ export function useState(handle: S3HandleBase): EditorState {
   const redirect = useRedirect()
   const onSave = React.useCallback(async () => {
     setSaving(true)
-    const h = await writeFile(value || '') // TODO: Ask if user really wants to save empty file
-    setEditing(false)
-    setSaving(false)
-    redirect(h)
+    // XXX: implement custom MUI Dialog-based confirm?
+    // eslint-disable-next-line no-restricted-globals, no-alert
+    if (!value && !window.confirm('You are about to save empty file')) return
+    try {
+      setError(null)
+      const h = await writeFile(value || '')
+      setEditing(false)
+      setSaving(false)
+      redirect(h)
+    } catch (e) {
+      setError(e instanceof Error ? e : new Error(`${e}`))
+    }
   }, [redirect, value, writeFile])
   const onCancel = React.useCallback(() => setEditing(false), [])
   const onEdit = React.useCallback(() => setEditing(true), [])
   return React.useMemo(
     () => ({
       editing,
+      error,
       onCancel,
       onChange: setValue,
       onEdit,
@@ -65,23 +83,31 @@ export function useState(handle: S3HandleBase): EditorState {
       type,
       value,
     }),
-    [editing, onCancel, onEdit, onSave, saving, type, value],
+    [editing, error, onCancel, onEdit, onSave, saving, type, value],
   )
 }
 
 interface EditorProps {
   disabled?: boolean
   empty?: boolean
+  error: Error | null
   handle: S3HandleBase
   onChange: (value: string) => void
   type: EditorInputType
 }
 
-function EditorSuspended({ disabled, empty, handle, onChange, type }: EditorProps) {
+function EditorSuspended({
+  disabled,
+  empty,
+  error,
+  handle,
+  onChange,
+  type,
+}: EditorProps) {
   loadMode(type.brace || 'text')
 
   const data = PreviewUtils.useObjectGetter(handle, { noAutoFetch: empty })
-  if (empty) return <TextEditor type={type} value="" onChange={onChange} />
+  if (empty) return <TextEditor error={error} type={type} value="" onChange={onChange} />
   return data.case({
     _: () => <Skeleton />,
     Err: (
@@ -95,7 +121,13 @@ function EditorSuspended({ disabled, empty, handle, onChange, type }: EditorProp
     Ok: (response: $TSFixMe) => {
       const value = response.Body.toString('utf-8')
       return (
-        <TextEditor disabled={disabled} type={type} value={value} onChange={onChange} />
+        <TextEditor
+          disabled={disabled}
+          error={error}
+          onChange={onChange}
+          type={type}
+          value={value}
+        />
       )
     },
   })
